==> long-term/Process_steps/step3_sort_long_event.ipynb <==
import argparse
import os
import sys
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from datetime import datetime
import math
from IPython.display import display
from multiprocessing import cpu_count,Pool 
from joblib import Parallel, delayed
import psutil
from orderedset import OrderedSet

long_event_ = pd.read_hdf('../../long_events_.h5',key='LN_300')

subset_df =  long_event_[long_event_.Type != 'T']

subset_df2 = long_event_[long_event_.Type == 'T']

ne_LN = pd.concat([subset_df,subset_df2])

ne_LN

ne_LN.to_hdf('../../long_events_.h5',key='LN_300')
