==> long-term/Process_steps/step5_Chuck_events.ipynb <==
import argparse
import os
import sys
import psutil
import numpy as np
import pandas as pd
import matplotlib
matplotlib.use('Agg')
import matplotlib.pyplot as plt
#from datetime import datetime
import math
from IPython.display import display
from multiprocessing import cpu_count,Pool 
import multiprocessing
from joblib import Parallel, delayed
from orderedset import OrderedSet
import datetime

long_event_drop2 = pd.read_hdf('../../long_events_.h5',key='LN_300_nc2')

W_set = long_event_drop2[long_event_drop2.Type=='W']

data_split_W = np.array_split(W_set,4)

for i in range(0,len(data_split_W)):
    data_split_W[i].to_hdf('../../LG.h5',key='W_split_'+str(i))

T_set = long_event_drop2[long_event_drop2.Type!='W']

data_split_T = np.array_split(T_set,4)

for i in range(0,len(data_split_T)):
    data_split_T[i].to_hdf('../../LG.h5',key='T_split_'+str(i))

lad = pd.read_hdf('../../LG.h5',key='T_split_0')

lad
